Re-added tests to `test_pareto_front`
